No delete elsewhere

+++ probius.py
@@ -305,7 +305,7 @@ class MyClient(discord.Client):
 			await mainProbius(self,message,texts)
 		elif '[' in message.content:
 			await mainProbius(self,message,[message.content.split('[')[1].lower().split('/')])
-			if message.content[0]=='[':
+			if message.content[0]=='[' and message.guild.id in [535256944106012694,603924426769170433]:
 				await message.delete()
 		#if message.author.id==410481791204327424:
 			#await message.add_reaction('<:OrphAYAYA:657172520092565514>')
